fix(merkle_storage): increase ref_counts from u16 to u32 to avoid overflow

+++ storage/src/merkle_storage.rs
@@ -201,7 +201,7 @@ pub struct MerkleStorage {
     /// list of all actions done on staging area
     actions: Arc<Vec<Action>>,
     /// reference counts of entries for garbage collection
-    ref_counts: HashMap<EntryHash, u16>,
+    ref_counts: HashMap<EntryHash, u32>,
 }
 
 #[derive(Debug, Fail)]
